pyproject.toml
[project]
name = "langent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain==1.2.3",
    "langchain-community==0.4.1",
    "langchain-core==1.2.17",
    "langchain-experimental>=0.4.1",
    "langchain-openai==1.1.7",
    "langgraph==1.0.5",
    "python-dotenv>=1.2.2",
]
